Crate pd_client

Re-exports

pub use self::errors::Error;

pub use self::errors::Result;

Modules

client

config

errors

feature_gate

metrics

util

Structs

Config
The configuration for a PD Client.

DummyPdClient

Feature

FeatureGate

RegionInfo

RegionStat

RpcClient

Constants

INVALID_ID

RECONNECT_INTERVAL_SEC

REQUEST_TIMEOUT

Traits

PdClient
PdClient communicates with Placement Driver (PD). Because now one PD only supports one cluster, so it is no need to pass cluster id in trait interface every time, so passing the cluster id when creating the PdClient is enough and the PdClient will use this cluster id all the time.

Functions

take_peer_address
Takes the peer address (for sending raft messages) from a store.

validate_endpoints

Type Definitions

Key

PdFuture